backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.2.0
piexif==1.1.3
boto3==1.34.44
celery==5.3.6
redis==5.0.1
websockets==12.0
pytest==8.0.0
pytest-asyncio==0.23.5
ruff==0.2.2
mypy==1.8.0
httpx==0.26.0
python-multipart==0.0.9
